conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
aioredis=2.0.1=pyhd8ed1ab_0
anyio=3.5.0=py310hff52083_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h7f8727e_0
asgiref=3.5.0=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
bcrypt=3.2.0=py310h7f8727e_0
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.3.29=h06a4308_0
certifi=2021.5.30=py310h06a4308_0
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.2=py310hff52083_0
cryptography=36.0.0=py310h9ce1e76_0
ecdsa=0.17.0=pyhd3eb1b0_0
fastapi=0.75.1=pyhd8ed1ab_0
future=0.18.2=py310h06a4308_1
gmp=6.2.1=h2531618_2
gunicorn=20.1.0=py310hff52083_2
h11=0.13.0=pyhd8ed1ab_0
httptools=0.3.0=py310h6acc77f_1
idna=3.3=pyhd8ed1ab_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_15
libgomp=11.2.0=h1d223b6_15
libstdcxx-ng=9.3.0=hd4cf53a_17
libuuid=1.0.3=h7f8727e_2
libuv=1.43.0=h7f98852_0
ncurses=6.3=h7f8727e_2
openssl=1.1.1n=h7f8727e_0
orjson=3.6.7=py310h6acc77f_0
passlib=1.7.4=pyhd3eb1b0_0
pip=21.2.4=py310h06a4308_0
pyasn1=0.4.8=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pycryptodome=3.10.1=py310hdd9d17f_0
pydantic=1.9.0=py310h5764c6d_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py310hff52083_5
python=3.10.4=h12debd9_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python-jose=3.3.0=pyhd3eb1b0_0
python-multipart=0.0.5=py_0
python_abi=3.10=2_cp310
pyyaml=6.0=py310h5764c6d_4
readline=8.1.2=h7f8727e_1
redis-server=6.2.6=h9b69904_0
requests=2.27.1=pyhd8ed1ab_0
rsa=4.7.2=pyhd3eb1b0_1
setuptools=58.0.4=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py310hff52083_3
sqlite=3.38.2=hc218d9a_0
starlette=0.17.1=pyhd8ed1ab_0
tk=8.6.11=h1ccaba5_0
typing-extensions=4.1.1=hd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=pyhd8ed1ab_0
uvicorn=0.17.6=py310hff52083_0
uvicorn-standard=0.17.6=hff52083_0
uvloop=0.16.0=py310h6acc77f_1
watchgod=0.8.2=pyhd8ed1ab_0
websockets=10.1=py310h6acc77f_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7f98852_2
zlib=1.2.11=h7f8727e_4
